Spark 2.4 Pyspark Compatiblity
Several version bumps of associated libraries. Also includes
converting the current subprocess.py and associated PythonJob
to use an authentication token when starting Pyspark. The method
used for generating the token is the same as inside native Spark. The
token is passed via environment variable from the PythonJob class to
the subprocess.py process which starts up the python interpreter.